Move it/*_test.go to cmd package
