+++ .travis.yml
@@ -5,6 +5,5 @@ go:
   - tip
 
 script:
-  # Ignore ./it and ./internal/testutil
-  - go vet -v $(go list -f '{{ .ImportPath }}' github.com/vim-volt/volt/... | grep -v -E '/volt/(it|internal/testutil)$')
+  - go vet -v ./...
   - make test
