+++ cmd/profile_test.go
@@ -1,4 +1,4 @@
-package it
+package cmd
 
 import (
 	"testing"
